+++ property-lock/src/main/java/com/tianjian/property/management/service/RoomDoorService.java
@@ -5,7 +5,6 @@ import com.tianjian.property.bean.Door;
 import com.tianjian.property.bean.vo.DoorVo;
 import com.tianjian.property.utils.PageResult;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -15,7 +14,7 @@ public interface RoomDoorService {
     Map<String, List<DoorVo>> screenRoomDoor(Integer propertyid,Integer bulidingid ,String unitname,Integer pageNum,Integer pageSize);
     Map<String,String> selectdoorparticulars(Integer doorid);
     //模糊搜索
-    List<Door> fuzzySearch(Integer propertyid,String fuzzy, Integer pageNum, Integer pageSize);
+    List<Door> fuzzySearch(Integer propertyid, String fuzzy, Integer pageNum, Integer pageSize);
 
-    Map addDoor(List<Map> door, Double appUID) throws Exception;
+    Map addDoor(List<Map> door, Integer appUID) throws Exception;
 }
